Remake file structure of store
