--- project/src/store/notifications-slice/selectors.ts
@@ -1,5 +1,0 @@
-import { NameSpace } from '../../const';
-import { Notification } from '../../types/notification';
-import { State } from '../../types/store';
-
-export const getNotifications = (state: State): Notification[] => state[NameSpace.Notification].notifications;
