+++ project/src/store/reducers/favorites/favorites.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { NameSpace, Status } from '../../const';
-import { Offer } from '../../types/offer';
-import { fetchFavoritesAction, toggleFavoriteAction } from '../api-actions';
+import { NameSpace, Status } from '../../../const';
+import { Offer } from '../../../types/offer';
+import { fetchFavoritesAction, toggleFavoriteAction } from '../../api-actions';
 
 type InitialState = {
   favorites: Offer[];
