+++ project/src/store/reducers/favorites/selectors.ts
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { NameSpace, Status } from '../../const';
-import { Offer } from '../../types/offer';
-import { State } from '../../types/store';
+import { NameSpace, Status } from '../../../const';
+import { Offer } from '../../../types/offer';
+import { State } from '../../../types/store';
 
 export const getFavorites = (state: State): Offer[] => state[NameSpace.Favorites].favorites;
 export const getFetchStatus = (state: State): Status => state[NameSpace.Favorites].fetchStatus;
