+++ project/src/store/reducers/near-offers/selectors.ts
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { NameSpace, Status } from '../../const';
-import { Offer } from '../../types/offer';
-import { State } from '../../types/store';
+import { NameSpace, Status } from '../../../const';
+import { Offer } from '../../../types/offer';
+import { State } from '../../../types/store';
 
 export const getNearOffers = (state: State): Offer[] => state[NameSpace.NearOffers].nearOffers;
 export const getStatus = (state: State): Status => state[NameSpace.NearOffers].status;
