+++ project/src/store/reducers/offer/offer.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { NameSpace, Status } from '../../const';
-import { Offer } from '../../types/offer';
-import { fetchOfferAction, toggleFavoriteAction } from '../api-actions';
+import { NameSpace, Status } from '../../../const';
+import { Offer } from '../../../types/offer';
+import { fetchOfferAction, toggleFavoriteAction } from '../../api-actions';
 
 type InitialState = {
   offer: Offer | null;
