+++ project/src/store/reducers/offer/selectors.ts
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { NameSpace, Status } from '../../const';
-import { Offer } from '../../types/offer';
-import { State } from '../../types/store';
+import { NameSpace, Status } from '../../../const';
+import { Offer } from '../../../types/offer';
+import { State } from '../../../types/store';
 
 export const getOffer = (state: State): Offer | null => state[NameSpace.Offer].offer;
 export const getStatus = (state: State): Status => state[NameSpace.Offer].status;
