+++ project/src/store/reducers/offers/selectors.ts
@@ -1,7 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { NameSpace, Status } from '../../const';
-import { Offer } from '../../types/offer';
-import { State } from '../../types/store';
+import { NameSpace, Status } from '../../../const';
+import { Offer } from '../../../types/offer';
+import { State } from '../../../types/store';
 
 export const getOffers = (state: State): Offer[] => state[NameSpace.Data].offers;
 export const getStatus = (state: State): Status => state[NameSpace.Data].status;
